requirements-dev.txt
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
codecov
coverage
flake8
pytest
sphinx
# These are dependencies of various sphinx extensions for documentation.
ipython
matplotlib
numpydoc
sphinx-copybutton
sphinx_rtd_theme
sphinxcontrib-bibtex
